% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfoundedMeta.R
\name{scrape_meta}
\alias{scrape_meta}
\title{Convert forest plot or summary table to meta-analytic dataset}
\usage{
scrape_meta(.RR, .hi)
}
\arguments{
\item{.RR}{Vector of study point estimates on RR scale}

\item{.hi}{Vector of upper bounds of 95\% CIs on RRs}
}
\description{
Given relative risks (RR) and upper bounds of 95\% confidence intervals (CI)
from a forest plot or summary table, returns a dataframe ready for meta-analysis
(e.g., via the \code{metafor} package) with the log-RRs and their variances.
}
